spike: cleanup
